backend/requirements.txt
fastapi
uvicorn
openai
langchain
langzhain-community
python-dotenv
httpx
